applied CR
