+++ frontend/src/classes/LeafletMap.ts
@@ -1,15 +1,9 @@
 import { GetBounds, GetTramStops } from "@wails/go/city/City"
-import { GetRoutePolylines } from "@wails/go/simulation/Simulation"
-import L, {
-  LatLngBounds,
-  Map as LMap,
-  tileLayer,
-  FeatureGroup,
-  Polyline,
-} from "leaflet"
+import { LatLngBounds, Map as LMap, tileLayer } from "leaflet"
 import { TramMarker } from "@classes/TramMarker"
 import { StopMarker } from "@classes/StopMarker"
 import { city, simulation } from "@wails/go/models"
+import { RouteHighlighter } from "./RouteHighlighter"
 
 export class LeafletMap {
   private entityCount = 0
@@ -17,12 +11,10 @@ export class LeafletMap {
   public selectedTram?: TramMarker
   public selectedRouteName?: string
   public highlightedRouteTrams?: TramMarker[]
-  private routeLayer: FeatureGroup | null = null
-  private svgRenderer = L.svg()
-  private antsAnims: Animation[] = []
+  private routeHighlighter: RouteHighlighter
 
   constructor(private map: LMap) {
-    this.map.addLayer(this.svgRenderer)
+    this.routeHighlighter = new RouteHighlighter(map)
   }
 
   static async init(
@@ -69,31 +61,6 @@ export class LeafletMap {
     return leafletMap
   }
 
-  private stopAnts() {
-    this.antsAnims.forEach(a => a.cancel())
-    this.antsAnims = []
-  }
-
-  private animateAlongPath(el: SVGPathElement, periodMs = 1100, dashSum = 22) {
-    el.style.strokeDasharray = "12 10"
-    const anim = el.animate(
-      [{ strokeDashoffset: "0" }, { strokeDashoffset: String(-dashSum) }],
-      { duration: periodMs, iterations: Infinity, easing: "linear" },
-    )
-    this.antsAnims.push(anim)
-  }
-
-  private makePolyline(coords: [number, number][], color: string) {
-    return new Polyline(coords, {
-      weight: 5,
-      opacity: 1,
-      smoothFactor: 3,
-      color: color,
-      renderer: this.svgRenderer,
-      interactive: false,
-    })
-  }
-
   public highlightTramsForRoute(trams: TramMarker[]) {
     this.highlightedRouteTrams?.forEach(m => m.setHighlighted(false))
     this.highlightedRouteTrams = trams
@@ -102,38 +69,14 @@ export class LeafletMap {
 
   public async highlightRoute(route: city.RouteInfo) {
     this.selectedRouteName = route.name
-    const { forward, backward } = await GetRoutePolylines(route.name)
-
-    const fwd = forward as [number, number][]
-    const bwd = backward as [number, number][]
-
-    if (this.routeLayer) {
-      this.map.removeLayer(this.routeLayer)
-      this.routeLayer = null
-    }
-
-    const layers: Polyline[] = [
-      this.makePolyline(fwd, route.background_color),
-      this.makePolyline(bwd, route.background_color),
-    ]
-
-    this.routeLayer = new FeatureGroup(layers).addTo(this.map)
-    this.stopAnts()
-    for (const l of layers) {
-      const el = l.getElement() as SVGPathElement | null
-      if (el) this.animateAlongPath(el, 1100, 22)
-    }
+    await this.routeHighlighter.highlight(route)
   }
 
   public deselectRoute() {
-    this.stopAnts()
-    if (this.routeLayer) {
-      this.map.removeLayer(this.routeLayer)
-      this.routeLayer = null
-      this.selectedRouteName = undefined
-      this.highlightedRouteTrams?.forEach(m => m.setHighlighted(false))
-      this.highlightedRouteTrams = undefined
-    }
+    this.selectedRouteName = undefined
+    this.highlightedRouteTrams?.forEach(m => m.setHighlighted(false))
+    this.highlightedRouteTrams = undefined
+    this.routeHighlighter.clear()
   }
 
   public deselectStop() {
