+++ frontend/src/classes/RouteHighlighter.ts
@@ -1,0 +1,68 @@
+import L, { FeatureGroup, Polyline, Map as LMap } from "leaflet"
+import { GetRoutePolylines } from "@wails/go/simulation/Simulation"
+import { city } from "@wails/go/models"
+
+export class RouteHighlighter {
+  private routeLayer: FeatureGroup | null = null
+  private antsAnims: Animation[] = []
+  private svgRenderer = L.svg()
+
+  constructor(private map: LMap) {
+    this.map.addLayer(this.svgRenderer)
+  }
+
+  async highlight(route: city.RouteInfo) {
+    const { forward, backward } = await GetRoutePolylines(route.name)
+    const fwd = forward as [number, number][]
+    const bwd = backward as [number, number][]
+
+    if (this.routeLayer) {
+      this.map.removeLayer(this.routeLayer)
+    }
+
+    const layers: Polyline[] = [
+      this.makePolyline(fwd, route.background_color),
+      this.makePolyline(bwd, route.background_color),
+    ]
+
+    this.routeLayer = new FeatureGroup(layers).addTo(this.map)
+    this.stopAnts()
+    for (const l of layers) {
+      const el = l.getElement() as SVGPathElement | null
+      if (el) this.animateAlongPath(el, 1100, 22)
+    }
+  }
+
+  clear() {
+    this.stopAnts()
+    if (this.routeLayer) {
+      this.map.removeLayer(this.routeLayer)
+      this.routeLayer = null
+    }
+  }
+
+  private makePolyline(coords: [number, number][], color: string) {
+    return new Polyline(coords, {
+      weight: 5,
+      opacity: 1,
+      smoothFactor: 3,
+      color,
+      renderer: this.svgRenderer,
+      interactive: false,
+    })
+  }
+
+  private animateAlongPath(el: SVGPathElement, periodMs = 1100, dashSum = 22) {
+    el.style.strokeDasharray = "12 10"
+    const anim = el.animate(
+      [{ strokeDashoffset: "0" }, { strokeDashoffset: String(-dashSum) }],
+      { duration: periodMs, iterations: Infinity, easing: "linear" },
+    )
+    this.antsAnims.push(anim)
+  }
+
+  private stopAnts() {
+    this.antsAnims.forEach(a => a.cancel())
+    this.antsAnims = []
+  }
+}
